Move dgus_display to src/lcd
